@@ -1,32 +1,51 @@
 {
   "name": "wordpress/performance",
-  "type": "wordpress-plugin",
-  "license": "GPL-2.0-or-later",
   "description": "Performance plugin from the WordPress Performance Team, which is a collection of standalone performance features.",
-  "homepage": "https://wordpress.org/plugins/performance-lab/",
+  "license": "GPL-2.0-or-later",
+  "type": "wordpress-plugin",
   "keywords": [
     "performance",
     "wordpress"
   ],
+  "homepage": "https://wordpress.org/plugins/performance-lab/",
   "support": {
     "issues": "https://github.com/WordPress/performance/issues"
   },
-  "require-dev": {
-    "wp-phpunit/wp-phpunit": "^5.8",
-    "yoast/phpunit-polyfills": "^1.0"
-  },
   "require": {
-    "composer/installers": "~1.0",
-    "ext-dom": "*",
-    "ext-json": "*",
-    "php": ">=7|^8"
+    "php": "^7.2 || ^8.0",
+    "ext-json": "*"
+  },
+  "require-dev": {
+    "phpcompatibility/php-compatibility": "^9.3",
+    "phpstan/extension-installer": "^1.3",
+    "phpstan/phpstan": "^1.10",
+    "phpstan/phpstan-deprecation-rules": "^1.1",
+    "phpstan/phpstan-phpunit": "^1.3",
+    "slevomat/coding-standard": "^8.0",
+    "squizlabs/php_codesniffer": "^3.9",
+    "szepeviktor/phpstan-wordpress": "^1.3",
+    "wp-coding-standards/wpcs": "^3.1",
+    "wp-phpunit/wp-phpunit": "^6.5",
+    "yoast/phpunit-polyfills": "^1.0",
+    "phpstan/php-8-stubs": "^0.3.84"
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "PerformanceLab\\Tests\\": "tests/utils"
+    }
+  },
+  "config": {
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "phpstan/extension-installer": true
+    },
+    "platform": {
+      "php": "7.2"
+    }
   },
   "scripts": {
-    "post-install-cmd": "if php -r 'exit( version_compare( phpversion(), \"8.1\", \">=\" ) ? 0 : 1 );'; then composer --working-dir=build-cs install --no-interaction; else echo 'Skipping composer install for build-cs since not on PHP 8.1+. You are running: '; php -v;  fi",
-    "post-update-cmd": "if php -r 'exit( version_compare( phpversion(), \"8.1\", \">=\" ) ? 0 : 1 );'; then composer --working-dir=build-cs update --no-interaction; else echo 'Skipping composer update for build-cs since not on PHP 8.1+. You are running: '; php -v; fi",
-    "phpstan": "build-cs/vendor/bin/phpstan analyse --memory-limit=2048M",
-    "format": "build-cs/vendor/bin/phpcbf --report-summary --report-source",
-    "lint": "build-cs/vendor/bin/phpcs",
+    "format": "phpcbf --report-summary --report-source",
+    "lint": "phpcs",
     "lint:all": [
       "@lint",
       "@lint:auto-sizes",
