@@ -15,9 +15,9 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @since 1.0.0
  *
- * @return array
+ * @return array{label: string, status: string, badge: array{label: string, color: string}, description: string, actions: string, test: string} Result.
  */
-function perflab_aao_autoloaded_options_test() {
+function perflab_aao_autoloaded_options_test(): array {
 
 	$autoloaded_options_size  = perflab_aao_autoloaded_options_size();
 	$autoloaded_options_count = count( wp_load_alloptions() );
@@ -73,7 +73,7 @@ function perflab_aao_autoloaded_options_test() {
 	$result['description'] = apply_filters( 'perflab_aao_autoloaded_options_limit_description', $result['description'] );
 
 	$result['actions'] = sprintf(
-	/* translators: 1: HelpHub URL. 2: Link description. */
+		/* translators: 1: HelpHub URL. 2: Link description. */
 		'<p><a target="_blank" href="%1$s">%2$s</a></p>',
 		esc_url( __( 'https://wordpress.org/support/article/optimization/#autoloaded-options', 'performance-lab' ) ),
 		__( 'More info about performance optimization', 'performance-lab' )
@@ -99,9 +99,20 @@ function perflab_aao_autoloaded_options_test() {
  *
  * @return int autoloaded data in bytes.
  */
-function perflab_aao_autoloaded_options_size() {
+function perflab_aao_autoloaded_options_size(): int {
 	global $wpdb;
-	return (int) $wpdb->get_var( 'SELECT SUM(LENGTH(option_value)) FROM ' . $wpdb->prefix . 'options WHERE autoload = \'yes\'' );
+
+	$autoload_values = perflab_aao_get_autoload_values_to_autoload();
+
+	return (int) $wpdb->get_var(
+		$wpdb->prepare(
+			sprintf(
+				"SELECT SUM(LENGTH(option_value)) FROM $wpdb->options WHERE autoload IN (%s)",
+				implode( ',', array_fill( 0, count( $autoload_values ), '%s' ) )
+			),
+			$autoload_values
+		)
+	);
 }
 
 /**
@@ -111,9 +122,9 @@ function perflab_aao_autoloaded_options_size() {
  *
  * @global wpdb $wpdb WordPress database abstraction object.
  *
- * @return array Autoloaded data as option names and their sizes.
+ * @return array<string, object{option_name: string, option_value_length: int}> Autoloaded data as option names and their sizes.
  */
-function perflab_aao_query_autoloaded_options() {
+function perflab_aao_query_autoloaded_options(): array {
 	global $wpdb;
 
 	/**
@@ -130,7 +141,17 @@ function perflab_aao_query_autoloaded_options() {
 	 */
 	$option_threshold = apply_filters( 'perflab_aao_autoloaded_options_table_threshold', 100 );
 
-	return $wpdb->get_results( $wpdb->prepare( "SELECT option_name, LENGTH(option_value) AS option_value_length FROM {$wpdb->options} WHERE autoload='yes' AND LENGTH(option_value) > %d ORDER BY option_value_length DESC LIMIT 20", $option_threshold ) );
+	$autoload_values = perflab_aao_get_autoload_values_to_autoload();
+
+	return $wpdb->get_results(
+		$wpdb->prepare(
+			sprintf(
+				"SELECT option_name, LENGTH(option_value) AS option_value_length FROM {$wpdb->options} WHERE autoload IN (%s)",
+				implode( ',', array_fill( 0, count( $autoload_values ), '%s' ) )
+			) . ' AND LENGTH(option_value) > %d ORDER BY option_value_length DESC LIMIT 20',
+			array_merge( $autoload_values, array( $option_threshold ) )
+		)
+	);
 }
 
 /**
@@ -140,7 +161,7 @@ function perflab_aao_query_autoloaded_options() {
  *
  * @return string HTML formatted table.
  */
-function perflab_aao_get_autoloaded_options_table() {
+function perflab_aao_get_autoloaded_options_table(): string {
 	$autoload_summary = perflab_aao_query_autoloaded_options();
 
 	$html_table = sprintf(
@@ -174,13 +195,13 @@ function perflab_aao_get_autoloaded_options_table() {
 /**
  * Gets disabled autoload options table.
  *
- * @since n.e.x.t
+ * @since 3.0.0
  *
  * @global wpdb $wpdb WordPress database abstraction object.
  *
  * @return string HTML formatted table.
  */
-function perflab_aao_get_disabled_autoloaded_options_table() {
+function perflab_aao_get_disabled_autoloaded_options_table(): string {
 	global $wpdb;
 
 	$disabled_options = get_option( 'perflab_aao_disabled_options', array() );
@@ -226,4 +247,19 @@ function perflab_aao_get_disabled_autoloaded_options_table() {
 	$html_table .= '</tbody></table>';
 
 	return $html_table;
+}
+
+/**
+ * Gets the autoload values in the database that should trigger their option to be autoloaded.
+ *
+ * @since 3.0.0
+ *
+ * @return string[] List of autoload values.
+ */
+function perflab_aao_get_autoload_values_to_autoload(): array {
+	if ( function_exists( 'wp_autoload_values_to_autoload' ) ) {
+		return wp_autoload_values_to_autoload();
+	}
+
+	return array( 'yes' );
 }
