@@ -15,15 +15,18 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @since 1.0.0
  *
- * @return array
+ * @return array{label: string, status: string, badge: array{label: string, color: string}, description: string, actions: string, test: string}|array{omitted: true} Result.
  */
-function perflab_aea_enqueued_js_assets_test() {
+function perflab_aea_enqueued_js_assets_test(): array {
 	/**
 	 * If the test didn't run yet, deactivate.
 	 */
 	$enqueued_scripts = perflab_aea_get_total_enqueued_scripts();
 	if ( false === $enqueued_scripts ) {
-		return array();
+		// The return value is validated in JavaScript at:
+		// <https://github.com/WordPress/wordpress-develop/blob/d1e0a6241dcc34f4a5ed464a741116461a88d43b/src/js/_enqueues/admin/site-health.js#L65-L114>
+		// If the value lacks the required keys of test, label, and description then it is omitted.
+		return array( 'omitted' => true );
 	}
 
 	$result = array(
@@ -109,15 +112,16 @@ function perflab_aea_enqueued_js_assets_test() {
  *
  * @since 1.0.0
  *
- * @return array
+ * @return array{label: string, status: string, badge: array{label: string, color: string}, description: string, actions: string, test: string}|array{omitted: true} Result.
  */
-function perflab_aea_enqueued_css_assets_test() {
-	/**
-	 * If the test didn't run yet, deactivate.
-	 */
+function perflab_aea_enqueued_css_assets_test(): array {
+	// Omit if the test didn't run yet, omit.
 	$enqueued_styles = perflab_aea_get_total_enqueued_styles();
 	if ( false === $enqueued_styles ) {
-		return array();
+		// The return value is validated in JavaScript at:
+		// <https://github.com/WordPress/wordpress-develop/blob/d1e0a6241dcc34f4a5ed464a741116461a88d43b/src/js/_enqueues/admin/site-health.js#L65-L114>
+		// If the value lacks the required keys of test, label, and description then it is omitted.
+		return array( 'omitted' => true );
 	}
 	$result = array(
 		'label'       => __( 'Enqueued styles', 'performance-lab' ),
@@ -276,7 +280,7 @@ function perflab_aea_get_total_size_bytes_enqueued_styles() {
  * @param string $resource_url URl resource link.
  * @return string Returns absolute path to the resource.
  */
-function perflab_aea_get_path_from_resource_url( $resource_url ) {
+function perflab_aea_get_path_from_resource_url( string $resource_url ): string {
 	if ( ! $resource_url ) {
 		return '';
 	}
