@@ -75,6 +75,22 @@ function embed_optimizer_filter_oembed_html( string $html ): string {
 	if ( 1 === $iframe_count && $html_processor->has_bookmark( 'iframe' ) ) {
 		if ( $html_processor->seek( 'iframe' ) ) {
 			$html_processor->set_attribute( 'loading', 'lazy' );
+
+			// For post embeds, use visibility:hidden instead of clip since browsers will consistently load the
+			// lazy-loaded iframe (where Chromium is unreliably with clip) while at the same time improve accessibility
+			// by preventing links in the hidden iframe from receiving focus.
+			if ( $html_processor->has_class( 'wp-embedded-content' ) ) {
+				$style = $html_processor->get_attribute( 'style' );
+				if ( $style ) {
+					// WordPress core injects this clip CSS property:
+					// <https://github.com/WordPress/wordpress-develop/blob/6974b994de5/src/wp-includes/embed.php#L968>.
+					$style = str_replace( 'clip: rect(1px, 1px, 1px, 1px);', 'visibility: hidden;', $style );
+
+					// Note: wp-embed.js removes the style attribute entirely when the iframe is loaded:
+					// <https://github.com/WordPress/wordpress-develop/blob/6974b994d/src/js/_enqueues/wp/embed.js#L60>.
+					$html_processor->set_attribute( 'style', $style );
+				}
+			}
 		} else {
 			embed_optimizer_trigger_error( __FUNCTION__, esc_html__( 'Embed Optimizer unable to seek to iframe bookmark.', 'embed-optimizer' ) );
 		}
@@ -162,6 +178,7 @@ function embed_optimizer_trigger_error( string $function_name, string $message, 
  * @since 0.1.0
  */
 function embed_optimizer_render_generator() {
-	echo '<meta name="generator" content="Embed Optimizer ' . esc_attr( EMBED_OPTIMIZER_VERSION ) . '">' . "\n";
+	// Use the plugin slug as it is immutable.
+	echo '<meta name="generator" content="embed-optimizer ' . esc_attr( EMBED_OPTIMIZER_VERSION ) . '">' . "\n";
 }
 add_action( 'wp_head', 'embed_optimizer_render_generator' );
