@@ -71,13 +71,13 @@ function webp_uploads_get_upload_image_mime_transforms() {
  * @since 1.0.0
  * @access private
  *
- * @param int         $attachment_id         The ID of the attachment from where this image would be created.
- * @param string      $image_size            The size name that would be used to create the image source, out of the registered subsizes.
- * @param array       $size_data             An array with the dimensions of the image: height, width and crop.
- * @param string      $mime                  The target mime in which the image should be created.
- * @param string|null $destination_file_name The path where the file would be stored, including the extension. If null, `generate_filename` is used to create the destination file name.
+ * @param int                                          $attachment_id         The ID of the attachment from where this image would be created.
+ * @param string                                       $image_size            The size name that would be used to create the image source, out of the registered subsizes.
+ * @param array{ width: int, height: int, crop: bool } $size_data             An array with the dimensions of the image: height, width and crop.
+ * @param string                                       $mime                  The target mime in which the image should be created.
+ * @param string|null                                  $destination_file_name The path where the file would be stored, including the extension. If null, `generate_filename` is used to create the destination file name.
  *
- * @return array|WP_Error An array with the file and filesize if the image was created correctly, otherwise a WP_Error.
+ * @return array{ file: string, filesize: int }|WP_Error An array with the file and filesize if the image was created correctly, otherwise a WP_Error.
  */
 function webp_uploads_generate_additional_image_source( $attachment_id, $image_size, array $size_data, $mime, $destination_file_name = null ) {
 	/**
@@ -88,11 +88,19 @@ function webp_uploads_generate_additional_image_source( $attachment_id, $image_s
 	 *
 	 * @since 1.1.0
 	 *
-	 * @param array|null|WP_Error $image         Image data {'path'=>string, 'file'=>string, 'width'=>int, 'height'=>int, 'mime-type'=>string} or null or WP_Error.
-	 * @param int                 $attachment_id The ID of the attachment from where this image would be created.
-	 * @param string              $image_size    The size name that would be used to create this image, out of the registered subsizes.
-	 * @param array               $size_data     An array with the dimensions of the image: height, width and crop {'height'=>int, 'width'=>int, 'crop'}.
-	 * @param string              $mime          The target mime in which the image should be created.
+	 * @param array{
+	 *            file: string,
+	 *            path?: string,
+	 *            filesize?: int
+	 *        }|null|WP_Error $image         Image data, null, or WP_Error.
+	 * @param int             $attachment_id The ID of the attachment from where this image would be created.
+	 * @param string          $image_size    The size name that would be used to create this image, out of the registered subsizes.
+	 * @param array{
+	 *            width: int,
+	 *            height: int,
+	 *            crop: bool
+	 *        }               $size_data     An array with the dimensions of the image.
+	 * @param string          $mime          The target mime in which the image should be created.
 	 */
 	$image = apply_filters( 'webp_uploads_pre_generate_additional_image_source', null, $attachment_id, $image_size, $size_data, $mime );
 	if ( is_wp_error( $image ) ) {
@@ -157,7 +165,7 @@ function webp_uploads_generate_additional_image_source( $attachment_id, $image_s
 		$destination_file_name = $editor->generate_filename( $suffix, null, $extension[0] );
 	}
 
-	remove_filter( 'image_editor_output_format', 'webp_uploads_filter_image_editor_output_format', 10, 3 );
+	remove_filter( 'image_editor_output_format', 'webp_uploads_filter_image_editor_output_format', 10 );
 	$image = $editor->save( $destination_file_name, $mime );
 	add_filter( 'image_editor_output_format', 'webp_uploads_filter_image_editor_output_format', 10, 3 );
 
@@ -188,7 +196,7 @@ function webp_uploads_generate_additional_image_source( $attachment_id, $image_s
  * @param string $size          The size name that would be used to create this image, out of the registered subsizes.
  * @param string $mime          A mime type we are looking to use to create this image.
  *
- * @return array|WP_Error
+ * @return array{ file: string, filesize: int }|WP_Error
  */
 function webp_uploads_generate_image_size( $attachment_id, $size, $mime ) {
 	$sizes    = wp_get_registered_image_subsizes();
