@@ -234,7 +234,7 @@ class Dominant_Color_Test extends DominantColorTestCase {
 	public function test_dominant_color_update_attachment_image_attributes( $style_attr, $expected ) {
 		$attachment_id = self::factory()->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/plugins/dominant-color-images/data/images/red.jpg' );
 
-		$attachment_image = wp_get_attachment_image( $attachment_id, 'full', '', array( 'style' => $style_attr ) );
+		$attachment_image = wp_get_attachment_image( $attachment_id, 'full', false, array( 'style' => $style_attr ) );
 		$this->assertStringContainsString( $expected, $attachment_image );
 	}
 
@@ -370,5 +370,17 @@ class Dominant_Color_Test extends DominantColorTestCase {
 				'hex'   => null,
 			),
 		);
+	}
+
+	/**
+	 * Test printing the meta generator tag.
+	 *
+	 * @covers ::dominant_color_render_generator
+	 */
+	public function test_dominant_color_render_generator() {
+		$tag = get_echo( 'dominant_color_render_generator' );
+		$this->assertStringStartsWith( '<meta', $tag );
+		$this->assertStringContainsString( 'generator', $tag );
+		$this->assertStringContainsString( 'dominant-color-images ' . DOMINANT_COLOR_IMAGES_VERSION, $tag );
 	}
 }
