@@ -12,7 +12,7 @@ class OD_Hooks_Tests extends WP_UnitTestCase {
 	 *
 	 * @see OD_Storage_Post_Type_Tests::test_add_hooks()
 	 */
-	public function test_hooks_added() {
+	public function test_hooks_added(): void {
 		$this->assertEquals( PHP_INT_MAX, has_filter( 'template_include', 'od_buffer_output' ) );
 		$this->assertEquals( 10, has_filter( 'wp', 'od_maybe_add_template_output_buffer_filter' ) );
 		$this->assertSame(
